Handle file under-read
In sysfs all files are reported as 4K in size but provide a variable
size based on real content, in those cases we read a lot of zeroes
and wasted some space and made things less conventient to work with.

Now we adjust the total file size if the first read was below the first
buffer size.