Improve ordering of output messages
By taking the lock before outputing the message we get the message just
when we start to send the data for it out and not just when we are
ready.

It mostly makes a little more sense even though it probably reduces
performance a bit by "blocking" for logging while the lock is taken.